% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_model.R
\name{ai}
\alias{ai}
\title{Estimation of an Almost-Ideal Demand System}
\usage{
ai(w, p, x, z, a0 = 0, data, scale = FALSE, logp = TRUE, logexp = TRUE)
}
\arguments{
\item{w}{character vector of m budgetshares used for estimation.}

\item{p}{character vector of m prices.}

\item{x}{single character vector of total expenditure.}

\item{a0}{start value for translog price index.}

\item{data}{data.frame containing the variables.}

\item{scale}{logical if TRUE Rays (1983) scaling is used.}

\item{logp}{logical if prices are log prices.}

\item{logx}{logical if expenditure is log expenditure.}
}
\description{
Estimation of an Almost-Ideal Demand System using nlsur().
}
\references{
Deaton, Angus S., Muellbauer, John: An Almost Ideal Demand
 System, The American Economic Review 70(3), American Economic Association,
 312–326, 1980

Ray, Ranjan: Measuring the costs of children: An alternative
 approach, Journal of Public Economics 22(1), 89–102, 1983
}
\seealso{
qai and ai.model
}
