% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_reg}
\alias{calc_reg}
\title{calc_reg}
\usage{
calc_reg(x, r, qS, w, sizetheta, fullreg, tol)
}
\arguments{
\item{x}{x}

\item{r}{r}

\item{qS}{qS}

\item{w}{w}

\item{sizetheta}{sizetheta}

\item{fullreg}{fullreg}

\item{tol}{tol}
}
\description{
calc_reg
}
