% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_robust}
\alias{calc_robust}
\title{calc_robust}
\usage{
calc_robust(x, u, qS, w, sizetheta)
}
\arguments{
\item{x}{x}

\item{u}{u}

\item{qS}{qS}

\item{w}{w}

\item{sizetheta}{sizetheta}
}
\description{
calc_robust
}

