% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_robust}
\alias{cov_robust}
\title{cov_robust}
\usage{
cov_robust(x, u, qS, w, sizetheta)
}
\arguments{
\item{x}{matrix of derivatives}

\item{u}{u}

\item{qS}{weighting matrix}

\item{w}{vector of weights}

\item{sizetheta}{sizetheta}
}
\description{
As discussed in Wooldridge (2002, 160)
}
