% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcom.R
\name{dm}
\alias{dm}
\title{dm simple delta method implementation}
\usage{
dm(object, form, level = 0.05)
}
\arguments{
\item{object}{of class nlsur}

\item{form}{formula e.g. "be/bk".}

\item{level}{value for conf. interval default is 0.05}
}
\description{
dm simple delta method implementation
}
