% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_model.R
\name{elasQAI}
\alias{elasQAI}
\title{Estimation of elasticies of the Quadratic Almost-Ideal Demand System}
\usage{
elasQAI(object, data, type = 1, usemean = FALSE)
}
\arguments{
\item{object}{qai result}

\item{data}{data vector used for estimation}

\item{type}{1 = expenditure; 2 = uncompensated; 3 = compensated}

\item{usemean}{evaluate at mean}
}
\description{
Estimates the income/expeditute elasticity, the uncompensated price
 elasticity and the compensated price elasticity
}
\references{
Banks, James, Blundell, Richard, Lewbel, Arthur: Quadratic Engel
 Curves and Consumer Demand, The Review of Economics and Statistics 79(4),
 The MIT Press, 527-539, 1997

Poi, Brian P.: Easy demand-system estimation with quaids, The
 Stata Journal 12(3), 433-446, 2012
}
\seealso{
ai and ai.model
}
