% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_model.R
\name{elasticities}
\alias{elasticities}
\title{Estimation of elasticies of the (Quadratic) Almost-Ideal Demand System}
\usage{
elasticities(object, data, type = 1, usemean = FALSE)
}
\arguments{
\item{object}{qai result}

\item{data}{data vector used for estimation}

\item{type}{1 = expenditure; 2 = uncompensated; 3 = compensated}

\item{usemean}{evaluate at mean}
}
\description{
Estimates the income/expeditute elasticity, the uncompensated price
 elasticity and the compensated price elasticity
}
\details{
Formula for the expenditure (income) elasticity

\deqn{
\mu_i = 1 + \frac{1}{w_i} \left[ \beta_i + \frac{2\lambda_i}{b(\mathbf{p})} *
 ln \left\{\frac{m}{a(\mathbf{p})} \right\}\right]
}

Formula for the uncompensated price elasticity

\deqn{
\epsilon_{ij} = \delta_{ij} + \frac{1}{w_i} \left( \gamma_{ij} - \beta_i +
 \frac{2\lambda_i}{b(\mathbf{p})} \right) \left[\ln \left\{
 \frac{m}{a(\mathbf{p})}\right\} \right] \times \\
 \left(\alpha_j + \sum_k \gamma_{jk} \ln p_k \right) -
 \frac{\beta_j \lambda_i}{b(\mathbf{p})} \left[
  \ln \left\{ \frac{m}{a(\mathbf{p})} \right\}\right]
}

Compensated price elasticitys (Slutsky equation)
\deqn{
 \epsilon_{ij}^{C} = \epsilon_{ij} + \mu_i w_j
}
}
\examples{
\dontrun{
library(nlsur)
library(readstata13)

dd <- read.dta13("http://www.stata-press.com/data/r15/food.dta")

w <- paste0("w", 1:4); p <- paste0("p", 1:4); x <- "expfd"

est <- ai(w = w, p = p, x = x, data = dd, a0 = 10, scale = FALSE,
          logp = F, logexp = F)

mu <- elasticities(est, data = dd, type = 1, usemean = FALSE)

ue <- elasticities(est, data = dd, type = 2, usemean = FALSE)

ce <- elasticities(est, data = dd, type = 3, usemean = FALSE)}

}
\references{
Banks, James, Blundell, Richard, Lewbel, Arthur: Quadratic Engel
 Curves and Consumer Demand, The Review of Economics and Statistics 79(4),
 The MIT Press, 527-539, 1997

Poi, Brian P.: Easy demand-system estimation with quaids, The
 Stata Journal 12(3), 433-446, 2012
}
\seealso{
ai and qai
}
