% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_model.R
\name{eQAI}
\alias{eQAI}
\title{Estimation of elasticies of the Quadratic Almost-Ideal Demand System}
\usage{
eQAI(object, data, ...)
}
\arguments{
\item{object}{qai result}

\item{data}{data vector used for estimation}

\item{...}{additional options passed to function}
}
\description{
Estimates the expenditure elasticity for goods
}
\references{
Poi, Brian P.: Easy demand-system estimation with quaids, The
 Stata Journal 12(3), 433-446, 2012
}
\seealso{
ai and ai.model
}
