% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getstartvals}
\alias{getstartvals}
\title{Function to create startvalues for nlsur models}
\usage{
getstartvals(model, data, val)
}
\arguments{
\item{model}{nlsur model}

\item{data}{the data frame used for evaluation}

\item{val}{value}
}
\description{
Function to create startvalues for nlsur models
}

