% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcom.R
\name{nlcom}
\alias{nlcom}
\title{Estimate nonlinear combinations of nlsur estimates}
\usage{
nlcom(object, form, alpha = 0.05, rname, envir)
}
\arguments{
\item{object}{of class nlsur}

\item{form}{formula e.g. "be/bk". May contain names(coef(object)) or prior
nlcom estimations.}

\item{alpha}{value for conf. interval default is 0.05}

\item{rname}{optional rowname for result}

\item{envir}{optional name of environment to search for additional parameters}
}
\description{
Estimate nonlinear combinations of nlsur estimates
}
\examples{
\dontrun{
dkm <- nlcom(object = erg, form = "-dkk -dkl -dke", rname = "dkm")
dkm

dlm <- nlcom(object = erg, form = "-dkl -dll -dle", rname = "dlm")
dlm

dem <- nlcom(object = erg, form = "-dke -dle -dee", rname = "dem")
dem

dmm <- nlcom(object = erg, form = "-dkm -dlm -dem", rname = "dmm")
dmm

# last one is equivalent to the longer form of:
dmm <- nlcom(object = erg,
 form = "-(-dkk -dkl -dke) -(-dkl -dll -dle) -(-dke -dle -dee)")
dmm
}

}
\seealso{
\link{deltaMethod}
}
