% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsur.R
\name{print.summary.nlsur}
\alias{print.summary.nlsur}
\title{Print nlsur summary}
\usage{
print.summary.nlsur(x, digits, ...)
}
\arguments{
\item{x}{summary object of class nlsur}
}
\description{
Print nlsur summary
}
