% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_model.R
\name{qai}
\alias{qai}
\title{Estimation of an Quadratic Almost-Ideal Demand System}
\usage{
qai(w, p, x, z, a0 = 0, data, scale = FALSE, logp = TRUE,
  logexp = TRUE, ...)
}
\arguments{
\item{w}{character vector of m budgetshares used for estimation.}

\item{p}{character vector of m prices.}

\item{x}{single character vector of total expenditure.}

\item{z}{character vector of k demographic variables.}

\item{a0}{start value for translog price index.}

\item{data}{data.frame containing the variables.}

\item{scale}{logical if TRUE Rays (1983) scaling is used.}

\item{logp}{logical if prices are log prices.}

\item{logexp}{logical if expenditure is log expenditure.}

\item{...}{additional options passed to nlsur}
}
\description{
Estimation of an Quadratic Almost-Ideal Demand System using nlsur().
}
\references{
Banks, James, Blundell, Richard, Lewbel, Arthur: Quadratic Engel
 Curves and Consumer Demand, The Review of Economics and Statistics 79(4),
 The MIT Press, 527-539, 1997

Ray, Ranjan: Measuring the costs of children: An alternative
 approach, Journal of Public Economics 22(1), 89-102, 1983
}
\seealso{
ai and ai.model
}
