% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsur.R
\name{summary.nlsur}
\alias{summary.nlsur}
\title{Summary of nlsur objects}
\usage{
\method{summary}{nlsur}(object, noconst = TRUE, ...)
}
\arguments{
\item{object}{object of class nlsur.}

\item{noconst}{logical value determining if model uses a constant or not.}

\item{...}{additional parameters (currently not used)}
}
\description{
Summary of nlsur objects
}

