% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_robust}
\alias{cov_robust}
\title{calc_robust}
\usage{
cov_robust(x, u, qS, w, sizetheta)
}
\arguments{
\item{x}{x}

\item{u}{u}

\item{qS}{qS}

\item{w}{w}

\item{sizetheta}{sizetheta}
}
\description{
as reference see:
http://www.navipedia.net/index.php/Block-Wise_Weighted_Least_Square
}
