% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsur.R
\name{nlsur}
\alias{nlsur}
\title{Non-Linear Seemingly Unrelated Regression}
\usage{
nlsur(eqns, data, startvalues, S = NULL, debug = FALSE, nls = FALSE,
  fgnls = FALSE, ifgnls = FALSE, MASS = FALSE, trace = FALSE,
  eps = eps, tau = tau)
}
\arguments{
\item{eqns}{is can be a single equation or a quation system. If eqns is a
single equation it will internaly be converted to a list. Estimation of a
single equation might as well be done using \code{nls()}.}

\item{data}{is the data set on which the equation is applied. This can be of
every type \code{eval()} can handle.}

\item{startvalues}{is a vector of intial start values. For}

\item{S}{is a weighing matrix used for estimation in Feasible Generalized
Non-Linear Least Squares (FGNLS) and Iterative FGNLS. For \code{nlsur()}
this is assumed to be the identity matrix. Hence, it is not included. If
included S is expected to be a matrix.}

\item{debug}{is a logical if debug output should be included. This can be a
lot and might not be of any help to you.}

\item{nls}{is a logical and default if estimation is done for NLSUR or NLS.}

\item{fgnls}{is a logical and must be set, if estimation is done for FGNLS.
This is called in a function called \code{fgnls()} and should not be set by
the user.}

\item{ifgnls}{is a logical and must be set, if estimation is done for ifgnls.
This is called in a function called \code{ifgnls()} and should not be set by
the user.}

\item{MASS}{is a logical, if TRUE \code{lm.gls()} is called for estimation of
a linear regression with a weighting matrix. Otherwise Rs matrix functions
will be used. Estimation results with MASS might be more stable, but it is
unable to handle a \code{Matrix::Diagonal()}. This will be converted with
\code{as.matrix()} which can be quite RAM consuming.}

\item{trace}{is a logical. If TRUE the current iterations SSR is called.}

\item{eps}{the epislon used for convergence in nlsur(). Default is 1e-5.}

\item{tau}{is another convergence variable. Default is 1e-3.}
}
\description{
\code{nlsur()} is a function for estimation of a non-linear seemingly
 unrelated regression model in R.
}
\details{
nlsur is a function for estimation of a non-linear least squares
(NLS). In addition to \code{nls()} it is capable of estimation of system of
equations. This estimation is done in a non-linear seemingly unrelated
regression approach.
}
\references{
Bates, D. M. and Watts, D. G. (1988) Nonlinear Regression Analysis and Its
 Applications, Wiley

Gallant, A. Ronald (1987): Nonlinear Statistical Models. Wiley: New York
}

