% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsur.R
\name{predict.nlsur}
\alias{predict.nlsur}
\title{Predict for Non-Linear Seemingly Unrelated Regression Models}
\usage{
\method{predict}{nlsur}(object, newdata, ...)
}
\arguments{
\item{object}{is an nlsur estimation result.}

\item{newdata}{an optional data frame for which the prediction is evaluated.}

\item{...}{further arguments for predict. At present no optional arguments
are used.}
}
\description{
\code{predict()} is a function to predict nlsur results.
}
\details{
predict.nlsur evaluates the nlsur equation(s) given nlsurs
estimated parameters using either the original data.frame or newdata. Since
nlsur() restricts the data object only to complete cases observations with
missings will not be fitted.
}
\examples{
# predict(nlsurObj, dataframe)
}

