% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsur.R
\name{predict.nlsur}
\alias{predict.nlsur}
\title{Predict for Non-Linear Seemingly Unrelated Regression Models}
\usage{
\method{predict}{nlsur}(object, newdata, ...)
}
\arguments{
\item{object}{is an nlsur estimation result.}

\item{newdata}{an optional data frame for which the prediction is evaluated.}

\item{...}{further arguments for predict. At present no optional arguments
are used.}
}
\description{
\code{predict()} is a function to predict nlsur results.
}
\details{
In contrast to other regression objects nlsur does not store the
complete model in the resulting object. This requires a data object for
predict. A limitation due to the fact that nlsur estimations tend to be
estimated on larger data.frames.
}
\examples{
# predict(nlsurObj, dataframe)

}

