% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wls_est}
\alias{wls_est}
\title{wls_est}
\usage{
wls_est(x, r, qS, w, sizetheta, fullreg, tol)
}
\arguments{
\item{x}{matrix of derivatives}

\item{r}{residual matrix}

\item{qS}{weighting matrix of sizetheta x sizetheta}

\item{w}{vector of weights}

\item{sizetheta}{integer defining the amount of coefficients}

\item{fullreg}{bool defining if WLS or Cov is calculated}

\item{tol}{tolerance used for qr()}
}
\description{
as reference see:
http://www.navipedia.net/index.php/Block-Wise_Weighted_Least_Square
}
