% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wls_est}
\alias{wls_est}
\title{calc_reg}
\usage{
wls_est(x, r, qS, w, sizetheta, fullreg, tol)
}
\arguments{
\item{x}{x}

\item{r}{r}

\item{qS}{qS}

\item{w}{w}

\item{sizetheta}{sizetheta}

\item{fullreg}{fullreg}

\item{tol}{tol}
}
\description{
as reference see:
http://www.navipedia.net/index.php/Block-Wise_Weighted_Least_Square
}
